+++ src/index.ts
@@ -1,4 +1,5 @@
 import FlipCardPanel from './FlipCardPanel';
 export * from './types';
 export { default as FlipCard } from './FlipCard';
+export { default as OdometerCard } from './OdometerCard';
 export default FlipCardPanel;